chore: apply renames as in #4085
